initial commit: move data fetching logic to canonical
