+++ src/app/pages/FrontPage/index.jsx
@@ -26,6 +26,7 @@ import { ServiceContext } from '#contexts/ServiceContext';
 import FrontPageSection from '#containers/FrontPageSection';
 import MetadataContainer from '#containers/Metadata';
 import MostReadContainer from '#containers/MostRead';
+import RadioScheduleContainer from '#containers/RadioSchedule';
 import LinkedData from '#containers/LinkedData';
 import ATIAnalytics from '#containers/ATIAnalytics';
 import ChartbeatAnalytics from '#containers/ChartbeatAnalytics';
@@ -115,6 +116,7 @@ const FrontPage = ({ pageData, mostReadEndpointOverride }) => {
               maxTwoColumns
             />
           )}
+          <RadioScheduleContainer endpoint="/korean/bbc_korean_radio/radioschedule.json" />
         </StyledFrontPageDiv>
       </StyledFrontPageMain>
     </>
